services/notebook-processor/requirements.txt
aiofiles~=24.1.0
nats-py~=2.8.0
jinja2~=3.1.4
jupytext~=1.16.3
nbformat~=5.10.4
